+++ CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to `Jsend` will be documented in this file.
 
 ### Added
 
-- JSend\Exception as an Exception marker of all the package Exception
+- `JSend\Exception` as an Exception marker of all the package Exception
+- `JSend\JSend::fromJSON` to create an object from a JSON string or a `JsonSerializable` object
+- `JSend\JSend::fromArray` to replace `JSend\JSend::createFromArray`
 
 ### Fixed
 
@@ -19,7 +21,10 @@ All notable changes to `Jsend` will be documented in this file.
 ### Removed
 
 - Remove PHP5 support
-- Relying on SPL Exception
+- Relying on PHP SPL Exception
+- `JSend::createFromString`
+- `JSend::createFromArray`
+- accessing the constructor you must use one of the named constructors
 
 ## 1.2.3 - 2016-09-15